+++ src/main/java/org/mustangproject/ZUGFeRD/ZUGFeRDImporter.java
@@ -46,6 +46,9 @@ import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
 import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
 import java.util.*;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -65,16 +68,16 @@ public class ZUGFeRDImporter {
 	 */
 	private byte[] rawXML = null;
 	private String xmpString = null; // XMP metadata
+	private Document document;
 
 	public ZUGFeRDImporter(String pdfFilename) {
-		try {
-			BufferedInputStream bis = new BufferedInputStream(new FileInputStream(pdfFilename));
+		
+		try (InputStream bis = Files.newInputStream(Paths.get(pdfFilename), StandardOpenOption.READ)) {
 			extractLowLevel(bis);
-			bis.close();
 		} catch (IOException e) {
 			Logger.getLogger(ZUGFeRDImporter.class.getName()).log(Level.SEVERE, null, e);
 			throw new ZUGFeRDExportException(e);
-		}
+		} 
 	}
 
 	public ZUGFeRDImporter(InputStream pdfStream) {
@@ -125,7 +128,10 @@ public class ZUGFeRDImporter {
 
 
 	private void extractFiles(Map<String, PDComplexFileSpecification> names) throws IOException {
-		for (String filename : names.keySet()) {
+		for (String alias : names.keySet()) {
+
+			PDComplexFileSpecification fileSpec = names.get(alias);
+			String filename=fileSpec.getFilename();
 			/**
 			 * currently (in the release candidate of version 1) only one attached file with
 			 * the name ZUGFeRD-invoice.xml is allowed
@@ -133,7 +139,6 @@ public class ZUGFeRDImporter {
 			if ((filename.equals("ZUGFeRD-invoice.xml") || (filename.equals("zugferd-invoice.xml")) || filename.equals("factur-x.xml"))) { //$NON-NLS-1$
 				containsMeta = true;
 
-				PDComplexFileSpecification fileSpec = names.get(filename);
 				PDEmbeddedFile embeddedFile = fileSpec.getEmbeddedFile();
 				// String embeddedFilename = filePath + filename;
 				// File file = new File(filePath + filename);
@@ -142,15 +147,13 @@ public class ZUGFeRDImporter {
 				// ByteArrayOutputStream();
 				// FileOutputStream fos = new FileOutputStream(file);
 
-				rawXML = embeddedFile.toByteArray();
-				setMeta(new String(rawXML));
+				setRawXML(embeddedFile.toByteArray());
 
 				// fos.write(embeddedFile.getByteArray());
 				// fos.close();
 			}
 			if (filename.startsWith("additional_data")) {
 
-				PDComplexFileSpecification fileSpec = names.get(filename);
 				PDEmbeddedFile embeddedFile = fileSpec.getEmbeddedFile();
 				additionalXMLs.put(filename, embeddedFile.toByteArray());
 
@@ -173,13 +176,52 @@ public class ZUGFeRDImporter {
 		System.err.println(output);
 	}
 
-	private Document getDocument() throws ParserConfigurationException, IOException, SAXException, TransformerException {
+		private Document getDocument() { return document; }
+
+	public void setRawXML(byte[] rawXML) throws IOException {
+		this.rawXML = rawXML;
+		try {
+			setDocument();
+		} catch (ParserConfigurationException | SAXException e) {
+			Logger.getLogger(ZUGFeRDImporter.class.getName()).log(Level.SEVERE, null, e);
+			throw new ZUGFeRDExportException(e);
+		}
+	}
+
+	private void setDocument() throws ParserConfigurationException, IOException, SAXException {
 		DocumentBuilderFactory xmlFact = DocumentBuilderFactory.newInstance();
 		xmlFact.setNamespaceAware(false);
 		DocumentBuilder builder = xmlFact.newDocumentBuilder();
-		Document doc = builder.parse(new ByteArrayInputStream(rawXML));
-		//prettyPrint(doc);
-		return doc;
+		ByteArrayInputStream is = new ByteArrayInputStream(rawXML);
+		is.skip(guessBOMSize(is));
+		document = builder.parse(is);
+	}
+
+	/**
+	 * Skips over a BOM at the beginning of the given ByteArrayInputStream, if one exists.
+	 * @param is the ByteArrayInputStream used
+	 * @throws IOException
+	 * @see <a href="https://www.w3.org/TR/xml/#sec-guessing">Autodetection of Character Encodings</a>
+	 */
+	private int guessBOMSize(ByteArrayInputStream is) throws IOException {
+		byte[] pad = new byte[4];
+		is.read(pad);
+		is.reset();
+		int test2 = ((pad[0] & 0xFF) << 8) | (pad[1] & 0xFF);
+		int test3 = ((test2 & 0xFFFF) << 8) | (pad[2] & 0xFF);
+		int test4 = ((test3 & 0xFFFFFF) << 8) | (pad[3] & 0xFF);
+		//
+		if (test4 == 0x0000FEFF || test4 == 0xFFFE0000 || test4 == 0x0000FFFE || test4 == 0xFEFF0000) {
+			// UCS-4: BOM takes 4 bytes
+			return 4;
+		} else if (test3 == 0xEFBBFF) {
+			// UTF-8: BOM takes 3 bytes
+			return 3;
+		} else if (test2 == 0xFEFF || test2 == 0xFFFE) {
+			// UTF-16: BOM takes 2 bytes
+			return 2;
+		}
+		return 0;
 	}
 
 	private String extractString(String xpathStr) {
@@ -192,10 +234,7 @@ public class ZUGFeRDImporter {
 			XPathFactory xpathFact = XPathFactory.newInstance();
 			XPath xpath = xpathFact.newXPath();
 			result = xpath.evaluate(xpathStr, document);
-		} catch (ParserConfigurationException e) {
-			Logger.getLogger(ZUGFeRDImporter.class.getName()).log(Level.SEVERE, null, e);
-			throw new ZUGFeRDExportException(e);
-		} catch (IOException | SAXException | TransformerException | XPathExpressionException e) {
+		} catch (XPathExpressionException e) {
 			Logger.getLogger(ZUGFeRDImporter.class.getName()).log(Level.SEVERE, null, e);
 			throw new ZUGFeRDExportException(e);
 		}
@@ -300,8 +339,8 @@ public class ZUGFeRDImporter {
 	/**
 	 * @param meta raw XML to be set
 	 */
-	public void setMeta(String meta) {
-		this.rawXML = meta.getBytes();
+	public void setMeta(String meta) throws IOException {
+		setRawXML(meta.getBytes());
 	}
 
 	/**
@@ -346,6 +385,7 @@ public class ZUGFeRDImporter {
 
 	/**
 	 * Returns the raw XML data as extracted from the ZUGFeRD PDF file.
+	 * @return the raw ZUGFeRD XML data
 	 */
 	public byte[] getRawXML() {
 		return rawXML;
@@ -369,7 +409,7 @@ public class ZUGFeRDImporter {
 	static String convertStreamToString(java.io.InputStream is) {
 		// source https://stackoverflow.com/questions/309424/how-do-i-read-convert-an-inputstream-into-a-string-in-java referring to
 		// https://community.oracle.com/blogs/pat/2004/10/23/stupid-scanner-tricks
-		Scanner s = new Scanner(is).useDelimiter("\\A");
+		Scanner s = new Scanner(is, "UTF-8").useDelimiter("\\A");
 		return s.hasNext() ? s.next() : "";
 	}