+++ src/test/java/org/mustangproject/ZUGFeRD/MustangReaderWriterTest.java
@@ -30,6 +30,7 @@ import org.apache.xmpbox.xml.DomXmpParser;
 import org.apache.xmpbox.xml.XmpParsingException;
 import org.junit.FixMethodOrder;
 import org.junit.runners.MethodSorters;
+import org.mustangproject.ZUGFeRD.model.DocumentContextParameterTypeConstants;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
@@ -184,15 +185,12 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 	// //////////////////////////////////////////////////////////////////////////////////////////
 
 	/**
-	 * The importer test imports from
-	 * ./src/test/MustangGnuaccountingBeispielRE-20170509_505.pdf to check the
-	 * values. --> as only Name Ascending is supported for Test Unit sequence, I
-	 * renamed the this test-A-Export to run before testZExport
-	 *
-	 * @throws IOException
+	 * The importer test imports from ./src/test/MustangGnuaccountingBeispielRE-20170509_505.pdf to check the
+	 * values. As only Name Ascending is supported for Test Unit sequence, I renamed the this testAImport
+	 * to run before testZExport
 	 */
 
-	public void testAImport() throws IOException {
+	public void testAImport() {
 		InputStream inputStream = this.getClass().getResourceAsStream("/MustangGnuaccountingBeispielRE-20170509_505.pdf");
 		ZUGFeRDImporter zi = new ZUGFeRDImporter(inputStream);
 
@@ -206,7 +204,7 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 		assertEquals(zi.getForeignReference(), "RE-20170509/505");
 	}
 
-	public void testForeignImport() throws IOException {
+	public void testForeignImport() {
 		InputStream inputStream = this.getClass().getResourceAsStream("/zugferd_invoice.pdf");
 		ZUGFeRDImporter zi = new ZUGFeRDImporter(inputStream);
 
@@ -343,16 +341,16 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 	 * ./src/test/MustangGnuaccountingBeispielRE-20140703_502blanko.pdf}, adds
 	 * metadata, writes to @{code ./target/testout-*} and then imports to check the
 	 * values. It would not make sense to have it run before the less complex
-	 * importer test (which is probably redundant) --> as only Name Ascending is
+	 * importer test (which is probably redundant). As only Name Ascending is
 	 * supported for Test Unit sequence, I renamed the Exporter Test test-Z-Export
 	 */
-	public void testZExport() throws Exception {
+	public void testZExport() {
 
 		final String TARGET_PDF = "./target/testout-MustangGnuaccountingBeispielRE-20171118_506new.pdf";
 
 		// the writing part
 		try (InputStream SOURCE_PDF = this.getClass()
-				.getResourceAsStream("/MustangGnuaccountingBeispielRE-20190610_507blanko.pdf");
+			.getResourceAsStream("/MustangGnuaccountingBeispielRE-20190610_507blanko.pdf");
 
 			 ZUGFeRDExporter ze = new ZUGFeRDExporterFromA1Factory().setZUGFeRDVersion(2).setZUGFeRDConformanceLevel(ZUGFeRDConformanceLevel.EN16931).load(SOURCE_PDF)) {
 
@@ -369,7 +367,9 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 //			assertFalse(pdfContent.indexOf("<zf:ConformanceLevel>EN 16931</zf:ConformanceLevel>") == -1);
 //			assertFalse(pdfContent.indexOf("<pdfaSchema:prefix>zf</pdfaSchema:prefix>") == -1);
 //			assertFalse(pdfContent.indexOf("urn:zugferd:pdfa:CrossIndustryDocument:invoice:2p0#") == -1);
-			
+
+		} catch (IOException e) {
+			fail("IOException should not happen in testZExport");
 		}
 
 		// now check the contents (like MustangReaderTest)
@@ -383,6 +383,37 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 		assertEquals(zi.getKTO(), getTradeSettlementPayment()[0].getOwnKto());
 		assertEquals(zi.getHolder(), getOwnOrganisationName());
 		assertEquals(zi.getForeignReference(), getNumber());
+	}
+
+	/**
+	 * Quick and dirty copy of testZExport to check if v1 files contain
+	 * the correct profile string when the comfort profile is selected.
+	 */
+	public void testZExportv1Profile() {
+
+		final String TARGET_PDF = "./target/testout-MustangGnuaccountingBeispielRE-20171118_506new.pdf";
+
+		// the writing part
+		try (InputStream SOURCE_PDF = this.getClass()
+			.getResourceAsStream("/MustangGnuaccountingBeispielRE-20190610_507blanko.pdf");
+
+			 ZUGFeRDExporter ze = new ZUGFeRDExporterFromA1Factory().setZUGFeRDVersion(1).setZUGFeRDConformanceLevel(ZUGFeRDConformanceLevel.COMFORT).load(SOURCE_PDF)) {
+
+			ze.PDFattachZugferdFile(this);
+			ze.disableAutoClose(true);
+			ze.export(TARGET_PDF);
+
+			ByteArrayOutputStream baos = new ByteArrayOutputStream();
+			ze.export(baos);
+			ze.close();
+			String pdfContent = baos.toString("UTF-8");
+			assertFalse(pdfContent.indexOf(DocumentContextParameterTypeConstants.BASIC) >= 0);
+			assertFalse(pdfContent.indexOf(DocumentContextParameterTypeConstants.EXTENDED) >= 0);
+			assertTrue(pdfContent.indexOf(DocumentContextParameterTypeConstants.COMFORT) >= 0);
+
+		} catch (IOException e) {
+			fail("IOException should not happen in testZExport");
+		}
 	}
 
 	
@@ -423,11 +454,7 @@ public class MustangReaderWriterTest extends MustangReaderTestCase {
 		assertEquals(zi.getForeignReference(), getNumber());
 	}
 
-	/**
-	 * @throws Exception
-	 * @Test(expected = IndexOutOfBoundsException.class)
-	 */
-	public void testExceptionOnPDF14() throws Exception {
+	public void testExceptionOnPDF14() {
 
 		final String TARGET_PDF = "./target/testout-MustangGnuaccountingBeispielRE-20170509_505new.pdf";